Code cleanup
git-svn-id: svn://scribus.net/branches/Version15x/Scribus@25419 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/ui/smfontfeatures.cpp
@@ -397,7 +397,7 @@ void SMFontFeatures::hideAllFontFeatures()
 
 void SMFontFeatures::enableFontFeatures(const QStringList& fontFeatures)
 {
-	int featureFlags = FontFeatures::fontFlagsFromList(fontFeatures);
+	quint64 featureFlags = FontFeatures::fontFlagsFromList(fontFeatures);
 
 	hideAllFontFeatures();
